Refactoring cache jobs helper methods into a separate concern
