+++ app/jobs/warm_up_quarterly_analytics_cache_job.rb
@@ -1,4 +1,5 @@
 class WarmUpQuarterlyAnalyticsCacheJob < ApplicationJob
+  include Cacheable
   queue_as :default
   self.queue_adapter = :sidekiq
 
@@ -8,8 +9,8 @@ class WarmUpQuarterlyAnalyticsCacheJob < ApplicationJob
       from_time_string = range[:from_time].strftime('%Y-%m-%d')
       to_time_string = range[:to_time].strftime('%Y-%m-%d')
 
-      ApplicationController.helpers.perform_facility_group_caching(from_time_string, to_time_string)
-      ApplicationController.helpers.perform_districts_caching(from_time_string, to_time_string)
+      perform_facility_group_caching(from_time_string, to_time_string)
+      perform_districts_caching(from_time_string, to_time_string)
     end
   end
 end
